Rename `RespawnListener` to `SpawnListener` and refactor logic.
Refactor to handle player spawn location logic within `SpawnListener` instead of `ConnectionListener`. This consolidates spawn-related handling and improves code organization.
