+++ src/main/java/net/thenextlvl/tweaks/listener/SpawnListener.java
@@ -7,10 +7,20 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerRespawnEvent;
+import org.spigotmc.event.player.PlayerSpawnLocationEvent;
 
 @RequiredArgsConstructor
-public class RespawnListener implements Listener {
+public class SpawnListener implements Listener {
     private final TweaksPlugin plugin;
+
+    @EventHandler(priority = EventPriority.LOWEST)
+    public void onPlayerSpawnLocation(PlayerSpawnLocationEvent event) {
+        var config = plugin.config().spawn();
+        if (config.location() == null) return;
+        if ((!config.teleportOnFirstJoin() || event.getPlayer().hasPlayedBefore())
+            && (!config.teleportOnJoin() || !event.getPlayer().hasPlayedBefore())) return;
+        event.setSpawnLocation(config.location());
+    }
 
     @EventHandler(priority = EventPriority.LOWEST)
     public void onPlayerRespawn(PlayerRespawnEvent event) {
